# advanced/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CineTrack Advanced API",
    "description": "CineTrack advanced-model API. The Flutter client talks only to this backend. Movie/TV discovery, details, cast/crew, seasons and episodes are fetched from TMDB, normalized and cached server-side. The API also provides PostgreSQL-backed users, JWT sessions, watch state/progress, ratings, spoiler-aware comments, favorites, personal lists, statistics, reports and role-based administration.",
    "contact": {
      "name": "CineTrack Project Team"
    },
    "license": {
      "name": "Educational non-commercial project"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Patch Profile",
        "operationId": "patch_profile_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/avatar": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Upload Avatar",
        "operationId": "upload_avatar_api_v1_users_me_avatar_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_v1_users_me_avatar_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete Avatar",
        "operationId": "delete_avatar_api_v1_users_me_avatar_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/stats": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Stats",
        "operationId": "get_stats_api_v1_users_me_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/activity": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Activity",
        "operationId": "get_activity_api_v1_users_me_activity_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityItem": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "media_title": {
            "type": "string",
            "title": "Media Title"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "media_id",
          "media_title",
          "media_type",
          "occurred_at"
        ],
        "title": "ActivityItem"
      },
      "ActivityResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActivityItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ActivityResponse"
      },
...
      "Body_upload_avatar_api_v1_users_me_avatar_post": {
        "properties": {
          "profile_image": {
            "type": "string",
            "format": "binary",
            "title": "Profile Image"
          }
        },
        "type": "object",
        "required": [
          "profile_image"
        ],
        "title": "Body_upload_avatar_api_v1_users_me_avatar_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileUpdateRequest": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40,
                "minLength": 3,
                "pattern": "^[A-Za-z0-9_.-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          }
        },
        "type": "object",
        "title": "ProfileUpdateRequest"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "username",
          "email",
          "avatar_url",
          "bio",
          "role",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse"
      },
      "UserStatsResponse": {
        "properties": {
          "watched_movies": {
            "type": "integer",
            "title": "Watched Movies"
          },
          "watched_series": {
            "type": "integer",
            "title": "Watched Series"
          },
          "watched_episodes": {
            "type": "integer",
            "title": "Watched Episodes"
          },
          "approximate_watch_minutes": {
            "type": "integer",
            "title": "Approximate Watch Minutes"
          },
          "approximate_watch_hours": {
            "type": "number",
            "title": "Approximate Watch Hours"
          },
          "favorite_genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favorite Genre"
          },
          "average_user_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average User Rating"
          },
          "favorites_count": {
            "type": "integer",
            "title": "Favorites Count"
          },
          "custom_lists_count": {
            "type": "integer",
            "title": "Custom Lists Count"
          }
        },
        "type": "object",
        "required": [
          "watched_movies",
          "watched_series",
          "watched_episodes",
          "approximate_watch_minutes",
          "approximate_watch_hours",
          "favorite_genre",
          "average_user_rating",
          "favorites_count",
          "custom_lists_count"
        ],
        "title": "UserStatsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}